requirements.txt
cloudpickle
Farama-Notifications
gymnasium
jax-jumpy
numpy
typing_extensions
